タイムラインのポストカードをコンポーネントにした・いいねをスムーズに
